Drop double from the monster sight chain (float-only thresholds)
Every idle Blake monster runs A_Look -> SightPlayer -> CheckSight ->
CheckSightTo each tic, and the chain was typed `double` even though the
DECORATE args are stored as `float` (ACTION_PARAM_DOUBLE expands to a float).
On the RV32 softcore (rv32imafc: hardware single-float, no double) every one
of those compares was a soft-double call.

Narrow the four sight-distance/fov params to `float` across the header and the
three definitions, and float-suffix the threshold literals (0.00001f, 359.75f,
179.5f, 180.5f) so the compares stay single-precision. Behaviour is identical:
the only thresholds in the whole actor set are {0, 0.1, 1, 180, 360}, all
exactly representable as float and none within a float ULP of a compare
boundary; distance is integer-valued and only ever tested against thresholds
<= 1. Callers already pass float locals or int literals.

+++ src/wolfenstein/port/wl_state.h
@@ -18,7 +18,7 @@ void SelectDodgeDir (AActor *ob);
 void SelectRunDir (AActor *ob);
 void SelectWanderDir (AActor *ob);
 bool MoveObj (AActor *ob, int32_t move);
-bool SightPlayer (AActor *ob, double minseedist, double maxseedist, double maxheardist, double fov, const Frame *state);
+bool SightPlayer (AActor *ob, float minseedist, float maxseedist, float maxheardist, float fov, const Frame *state);
 
 void    DamageActor (AActor *ob, AActor *attacker, unsigned damage);
